codex-tools: extract configured tool specs
