+++ codex-rs/tools/README.md
@@ -12,6 +12,7 @@ schema and Responses API tool primitives that no longer need to live in
 - `AdditionalProperties`
 - `ToolDefinition`
 - `ToolSpec`
+- `ConfiguredToolSpec`
 - `ResponsesApiTool`
 - `FreeformTool`
 - `FreeformToolFormat`
@@ -23,6 +24,7 @@ schema and Responses API tool primitives that no longer need to live in
 - `parse_tool_input_schema()`
 - `parse_dynamic_tool()`
 - `parse_mcp_tool()`
+- `create_tools_json_for_responses_api()`
 - `mcp_call_tool_result_output_schema()`
 - `tool_definition_to_responses_api_tool()`
 - `dynamic_tool_to_responses_api_tool()`
